Colocate buildAgentlessDefaultLayout with the other default builder (SCU-1713)
buildAgentlessDefaultLayout lived privately inside useWorkspaceShellBootstrap,
even though it is a pure sibling of buildDefaultWorkspaceLayout (it is defined by
stripping the placeholder center panel from that layout). Move it into
defaultLayout.ts and export it so both the bootstrap and the upcoming
reset-to-default-layout action build the arrangement from one place and cannot
drift. Pure relocation — no behavior change.
